feat(scene): Support "env" attribute in root scene
